requirements.txt
streamlit
python-dotenv
PyPDF2
transformers
torch